moved ViewEvent component to components folder and removed extra generateMetaData file
